backend/requirements.txt
python-dotenv
pyautogen
google-generativeai
torch
transformers
sentence-transformers
scikit-learn
numpy
pandas
joblib
networkx
matplotlib
fastapi
uvicorn
pydantic
plantuml
openai

